+++ src/NsisoLauncherCore/Util/Installer/Forge/ForgeInstaller.cs
@@ -1,15 +1,13 @@
-﻿using Newtonsoft.Json;
+﻿using System;
+using System.IO;
+using System.Threading;
+using System.Threading.Tasks;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using NsisoLauncherCore.Net;
 using NsisoLauncherCore.Net.Tools;
 using NsisoLauncherCore.Util.Installer.Forge.Actions;
 using NsisoLauncherCore.Util.Installer.Forge.Json;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace NsisoLauncherCore.Util.Installer.Forge
 {
@@ -25,30 +23,26 @@ namespace NsisoLauncherCore.Util.Installer.Forge
      */
     public class DataProcessorsForgeInstaller
     {
-        public string InstallerPath { get; set; }
-        public CommonInstallOptions Options { get; set; }
-
-        Install profile;
+        private Install profile;
 
         public DataProcessorsForgeInstaller(string installerPath, CommonInstallOptions options)
         {
             if (string.IsNullOrWhiteSpace(installerPath))
-            {
                 throw new ArgumentException("Installer path can not be null or whitespace.");
-            }
-            this.InstallerPath = installerPath;
-            this.Options = options ?? throw new ArgumentNullException("Install options is null");
+            InstallerPath = installerPath;
+            Options = options ?? throw new ArgumentNullException("Install options is null");
         }
 
-        public void BeginInstallFromJObject(ProgressCallback monitor, CancellationToken cancellationToken, JObject jObj, string tempPath)
+        public string InstallerPath { get; set; }
+        public CommonInstallOptions Options { get; set; }
+
+        public void BeginInstallFromJObject(ProgressCallback monitor, CancellationToken cancellationToken, JObject jObj,
+            string tempPath)
         {
             profile = jObj.ToObject<Install>();
 
-            string target = Options.GameRootPath;
-            if (!Directory.Exists(target))
-            {
-                throw new DirectoryNotFoundException("The minecraft root is not found");
-            }
+            var target = Options.GameRootPath;
+            if (!Directory.Exists(target)) throw new DirectoryNotFoundException("The minecraft root is not found");
 
             //I think we dont need to inject the launcher profiles, so we dont need this json :)
             //string launcherProfiles = Path.Combine(target, "launcher_profiles.json");
@@ -57,92 +51,73 @@ namespace NsisoLauncherCore.Util.Installer.Forge
             //    throw new FileNotFoundException("There is no minecraft launcher profile");
             //}
 
-            string versionRoot = Path.Combine(target, "versions");
-            string librariesDir = Path.Combine(target, "libraries");
-            if (!Directory.Exists(librariesDir))
-            {
-                Directory.CreateDirectory(librariesDir);
-            }
+            var versionRoot = Path.Combine(target, "versions");
+            var librariesDir = Path.Combine(target, "libraries");
+            if (!Directory.Exists(librariesDir)) Directory.CreateDirectory(librariesDir);
 
             //Extracting json
             monitor.SetState("提取json");
-            string jsonPath = PathManager.GetJsonPath(Options.GameRootPath, profile.Version);
+            var jsonPath = PathManager.GetJsonPath(Options.GameRootPath, profile.Version);
             if (!Directory.Exists(Path.GetDirectoryName(jsonPath)))
-            {
                 Directory.CreateDirectory(Path.GetDirectoryName(jsonPath));
-            }
             File.Copy(tempPath + profile.Json, jsonPath, true);
 
             //Consider minecraft client jar
             monitor.SetState("检查游戏文件");
-            string clientTarget = PathManager.GetJarPath(Options.GameRootPath, profile.Minecraft);
+            var clientTarget = PathManager.GetJarPath(Options.GameRootPath, profile.Minecraft);
             if (!File.Exists(PathManager.GetJsonPath(Options.GameRootPath, profile.Minecraft)))
-            {
                 throw new FileNotFoundException("Minecraft json is not exists");
-            }
-            if (!File.Exists(clientTarget))
-            {
-                throw new FileNotFoundException("Minecraft jar is not exists");
-            }
+            if (!File.Exists(clientTarget)) throw new FileNotFoundException("Minecraft jar is not exists");
 
-            var exc = DownloadUtils.DownloadForgeJLibraries(monitor, Options.Mirror, cancellationToken, profile.Libraries, librariesDir);
-            if (exc != null)
-            {
-                throw exc;
-            }
+            var exc = DownloadUtils.DownloadForgeJLibraries(monitor, Options.Mirror, cancellationToken,
+                profile.Libraries, librariesDir);
+            if (exc != null) throw exc;
 
-            string[] mavenFolders = Directory.GetDirectories(tempPath + "\\maven");
+            var mavenFolders = Directory.GetDirectories(tempPath + "\\maven");
             foreach (var item in mavenFolders)
             {
-                DirectoryInfo info = new DirectoryInfo(item);
+                var info = new DirectoryInfo(item);
                 FileHelper.CopyDirectory(item, librariesDir + '\\' + info.Name, true);
             }
 
-            PostProcessors postProcessors = new PostProcessors(profile, Options.IsClient, monitor);
-            Exception procExc = postProcessors.Process(tempPath, Options.GameRootPath, clientTarget,Options.Java);
-            if (procExc != null)
-            {
-                throw procExc;
-            }
+            var postProcessors = new PostProcessors(profile, Options.IsClient, monitor);
+            var procExc = postProcessors.Process(tempPath, Options.GameRootPath, clientTarget, Options.Java);
+            if (procExc != null) throw procExc;
         }
     }
 
     public class ForgeInstaller : IInstaller
     {
-        public string InstallerPath { get; set; }
-        public CommonInstallOptions Options { get; set; }
-
         public ForgeInstaller(string installerPath, CommonInstallOptions options)
         {
             if (string.IsNullOrWhiteSpace(installerPath))
-            {
                 throw new ArgumentException("Installer path can not be null or whitespace.");
-            }
-            this.InstallerPath = installerPath;
-            this.Options = options ?? throw new ArgumentNullException("Install options is null");
+            InstallerPath = installerPath;
+            Options = options ?? throw new ArgumentNullException("Install options is null");
         }
+
+        public string InstallerPath { get; set; }
+        public CommonInstallOptions Options { get; set; }
 
         public void BeginInstall(ProgressCallback monitor, CancellationToken cancellationToken)
         {
-            string installerName = Path.GetFileNameWithoutExtension(InstallerPath);
-            string tempPath = string.Format("{0}\\{1}Temp", PathManager.TempDirectory, installerName);
-            if (Directory.Exists(tempPath))
-            {
-                Directory.Delete(tempPath, true);
-            }
+            var installerName = Path.GetFileNameWithoutExtension(InstallerPath);
+            var tempPath = string.Format("{0}\\{1}Temp", PathManager.TempDirectory, installerName);
+            if (Directory.Exists(tempPath)) Directory.Delete(tempPath, true);
             Directory.CreateDirectory(tempPath);
             Unzip.UnZipFile(InstallerPath, tempPath);
-            string mainJson = File.ReadAllText(tempPath + "\\install_profile.json");
-            JObject jObject = JObject.Parse(mainJson);
+            var mainJson = File.ReadAllText(tempPath + "\\install_profile.json");
+            var jObject = JObject.Parse(mainJson);
 
             if (jObject.ContainsKey("install") && jObject.ContainsKey("versionInfo"))
             {
-                CommonInstaller commonInstaller = new CommonInstaller(InstallerPath, Options);
+                var commonInstaller = new CommonInstaller(InstallerPath, Options);
                 commonInstaller.BeginInstallFromJObject(monitor, cancellationToken, jObject, tempPath);
             }
-            else if (jObject.ContainsKey("data") && jObject.ContainsKey("processors") && jObject.ContainsKey("libraries"))
+            else if (jObject.ContainsKey("data") && jObject.ContainsKey("processors") &&
+                     jObject.ContainsKey("libraries"))
             {
-                DataProcessorsForgeInstaller forgeInstaller = new DataProcessorsForgeInstaller(InstallerPath, Options);
+                var forgeInstaller = new DataProcessorsForgeInstaller(InstallerPath, Options);
                 forgeInstaller.BeginInstallFromJObject(monitor, cancellationToken, jObject, tempPath);
             }
             else
@@ -156,10 +131,7 @@ namespace NsisoLauncherCore.Util.Installer.Forge
 
         public async Task BeginInstallAsync(ProgressCallback monitor, CancellationToken cancellationToken)
         {
-            await Task.Factory.StartNew(() =>
-            {
-                BeginInstall(monitor, cancellationToken);
-            });
+            await Task.Factory.StartNew(() => { BeginInstall(monitor, cancellationToken); });
         }
     }
 }
